+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # [NEXT]
 - Extensions improvements
+- `GameColorFilter` Improvements
 
 # [2.0.0-flamev1.2]
 - Fix bug in `getAnimation` (ImageExtension).